rrr: relocate MarshalSink/MarshalSource from serializable to marshal
Pure textual refactor (no module changes yet). MarshalSink,
MarshalSource, their adapters, and the make_*_proxy overloads moved
from serializable.hpp / serializable.cpp into marshal.hpp. The
member-function bodies (Sink::write / Source::read) become inline
in marshal.hpp since the full Marshal class def is in scope there.

BinaryWriteArchive(MarshalSink*) and BinaryReadArchive(MarshalSource*)
convenience constructors stay declared in serializable.hpp; their
bodies are now defined out-of-line as inline functions in
marshal.hpp (where MarshalSink/Source's full def is visible). This
preserves the public API (the autogenerated `benchmark_service.h`
and `client.hpp` use these conveniences directly).

Why: this breaks the impl-side cycle where serializable.cpp had to
include marshal.hpp to define MarshalSink::write/MarshalSource::read.
For the upcoming module conversion of serializable, that cycle was a
blocker — a module's exported-class member functions must live
inside the module, but including marshal.hpp from a module purview
gives Marshal module-attached linkage that conflicts with marshal.hpp's
global-module declaration when consumed textually elsewhere.

Build green; no behavior change.
